[android] fixed calculation of the animationDuration in the MyLocationView

+++ platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/maps/widgets/MyLocationView.java
@@ -997,7 +997,7 @@ public class MyLocationView extends View {
       }
 
       // updateLatLng timestamp
-      float previousUpdateTimeStamp = locationUpdateTimestamp;
+      long previousUpdateTimeStamp = locationUpdateTimestamp;
       locationUpdateTimestamp = SystemClock.elapsedRealtime();
 
       // calculate animation duration
@@ -1005,7 +1005,7 @@ public class MyLocationView extends View {
       if (previousUpdateTimeStamp == 0) {
         animationDuration = 0;
       } else {
-        animationDuration = (int) ((locationUpdateTimestamp - previousUpdateTimeStamp) * 1.1f)
+        animationDuration = (int) (((float)(locationUpdateTimestamp - previousUpdateTimeStamp)) * 1.1f)
         /*make animation slightly longer*/;
       }
